@@ -7,8 +7,8 @@
     "axios": "^0.27.2",
     "gh-pages": "^4.0.0",
     "leaflet": "^1.8.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-router": "^6.0.2",
     "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.1",
